Follow pattern bundleId.tests for naming unit test bundles
