+++ org.aniszczyk.minerva.tests.core/META-INF/MANIFEST.MF
@@ -2,7 +2,7 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Minerva
 Bundle-Vendor: aniszczyk.org
-Bundle-SymbolicName: org.aniszczyk.minerva.tests.core
+Bundle-SymbolicName: org.aniszczyk.minerva.core.tests
 Bundle-Version: 1.0.0.qualifier
 Require-Bundle: org.eclipse.core.runtime;bundle-version="[3.4.0,4.0.0)",
  org.eclipse.core.filesystem;bundle-version="[1.1.0,2.0.0)",
