+++ org.aniszczyk.minerva.tests.ui/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.aniszczyk.minerva.tests.ui</name>
+	<name>org.aniszczyk.minerva.ui.tests</name>
 	<comment></comment>
 	<projects>
 	</projects>
