+++ org.aniszczyk.minerva.tests.ui/META-INF/MANIFEST.MF
@@ -2,7 +2,7 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Minerva UI Tests
 Bundle-Vendor: aniszczyk.org
-Bundle-SymbolicName: org.aniszczyk.minerva.tests.ui
+Bundle-SymbolicName: org.aniszczyk.minerva.ui.tests
 Bundle-Version: 1.0.0.qualifier
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: J2SE-1.5
@@ -33,5 +33,5 @@ Import-Package: org.eclipse.swt.widgets,
  org.junit.runner;version="[4.3.1,5.0.0)",
  org.junit.runners;version="[4.3.1,5.0.0)",
  org.osgi.framework;version="[1.4.0,2.0.0)"
-Fragment-Host: org.aniszczyk.minerva.tests.core;bundle-version="1.0.0.qualifier"
-Export-Package: org.aniszczyk.minerva.tests.ui
+Fragment-Host: org.aniszczyk.minerva.core.tests;bundle-version="1.0.0.qualifier"
+Export-Package: org.aniszczyk.minerva.ui.tests
