+++ org.aniszczyk.minerva.tests.ui/src/org/aniszczyk/minerva/ui/tests/SimpleUITestCase.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.aniszczyk.minerva.tests.ui;
+package org.aniszczyk.minerva.ui.tests;
 
 import static org.junit.Assert.assertNotNull;
